HDX-9781 Move food_security_view
