+++ src/hapi_schema/db_views_as_tables.py
@@ -3,6 +3,7 @@ from datetime import datetime
 from sqlalchemy import (
     Boolean,
     DateTime,
+    Float,
     Integer,
     String,
 )
@@ -55,3 +56,31 @@ class DBDatasetVAT(Base):
     title: Mapped[str] = mapped_column(String(1024))
     hdx_provider_stub: Mapped[str] = mapped_column(String(128), index=True)
     hdx_provider_name: Mapped[str] = mapped_column(String(512), index=True)
+
+
+class DBFoodSecurityVAT(Base):
+    __tablename__ = "food_security_vat"
+    resource_hdx_id: Mapped[str] = mapped_column(String(36))
+    admin2_ref: Mapped[int] = mapped_column(Integer, primary_key=True)
+    ipc_phase: Mapped[str] = mapped_column(String(12), primary_key=True)
+    ipc_type: Mapped[str] = mapped_column(String(17), primary_key=True)
+    population_in_phase: Mapped[int] = mapped_column(Integer, index=True)
+    population_fraction_in_phase: Mapped[float] = mapped_column(
+        Float, index=True
+    )
+    reference_period_start: Mapped[datetime] = mapped_column(
+        DateTime, primary_key=True
+    )
+    reference_period_end: Mapped[datetime] = mapped_column(
+        DateTime, index=True
+    )
+    location_code: Mapped[str] = mapped_column(String(128), index=True)
+    location_name: Mapped[str] = mapped_column(String(512), index=True)
+    admin1_code: Mapped[str] = mapped_column(String(128), index=True)
+    admin1_name: Mapped[str] = mapped_column(String(512), index=True)
+    admin1_is_unspecified: Mapped[bool] = mapped_column(Boolean)
+    location_ref: Mapped[int] = mapped_column(Integer)
+    admin2_code: Mapped[str] = mapped_column(String(128), index=True)
+    admin2_name: Mapped[str] = mapped_column(String(512), index=True)
+    admin2_is_unspecified: Mapped[bool] = mapped_column(Boolean)
+    admin1_ref: Mapped[int] = mapped_column(Integer)
